+++ ext/torchaudio/extconf.rb
@@ -20,18 +20,9 @@ $CXXFLAGS += " -D_GLIBCXX_USE_CXX11_ABI=1"
 
 apple_clang = RbConfig::CONFIG["CC_VERSION_MESSAGE"] =~ /apple clang/i
 
-# check omp first
-if have_library("omp") || have_library("gomp")
-  $CXXFLAGS += " -Xclang" if apple_clang
-  $CXXFLAGS += " -fopenmp"
-end
-
 if apple_clang
-  # silence ruby/intern.h warning
-  $CXXFLAGS += " -Wno-deprecated-register"
-
   # silence torch warnings
-  $CXXFLAGS += " -Wno-shorten-64-to-32 -Wno-missing-noreturn"
+  $CXXFLAGS += " -Wno-deprecated-declarations"
 else
   # silence rice warnings
   $CXXFLAGS += " -Wno-noexcept-type"