+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.3.0 (unreleased)
+
+- Added `center`, `pad_mode`, and `onesided` options to `Spectogram` transform
+
 ## 0.2.1 (2021-07-16)
 
 - Added `create_dct` method